+++ utilities/reset.sh
@@ -1,0 +1,3 @@
+dropdb -U postgres georegistry
+createdb -U postgres -T template_postgis -O georegistry georegistry
+paster serve --reload development.ini