+++ CHANGES.rst
@@ -2,3 +2,4 @@
 ---
 - Ported code to Pyramid from Pylons
 - Ported database models
+- Ported GADM import script